Remove ui-component navbar and move into psss
